Schedule Trees (#1145)
This PR adds support for a scheduling-oriented view of SDFGs. Upon conversion, the SDFG and its nested SDFGs keep the same array names and are organized in one tree, where each node corresponds to a schedulable concept (map scope, copy, tasklet, for-loop scope, etc.). The graph structure can be converted to sequential text with `as_string`. Useful for inspecting and analyzing schedules. 

---------

Co-authored-by: Alexandros Nikolaos Ziogas <alexandros.ziogas@inf.ethz.ch>