Faster Call for `CompiledSDFG` (#1467)
This PR adds a new possibility how a `CompiledSDFG` can be called.

Before `__call__()` performed a lot of operations, which can be
summarized as:
- Ensuring that the arguments are ordered correctly.
- Transforming them to the right type (`ndarray` to pointers).

Especially when benchmarking smaller SDFG this is a lot of overhead,
which actually dominates the execution.
Furthermore, the runtime heavily depends on the number of arguments.

To solve this, this PR introduces the `_fast_call()` function, which
expects that its arguments are already in the right order and casted to
the right type.
In addition it does some refactoring and splits `_construct_args()` into
multiple parts.

---------

Co-authored-by: Philip Mueller <philip.paul.mueller@bluemain.ch>
Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>