Fixed `RedundantArray`'s handling of "reshaping" Memlets (#1603)
This PR fixes an issue that was reported as [issue
1595](https://github.com/spcl/dace/issues/1595), that was traced back to
`RedundantArray`.

The commit adds a deterministic test, unlike the one in the issue, that
fails without the fix.
The underlying problem is, that the transformation does not correctly
handle a Memlet that performs a reshaping.
This commit does not really solves the issue, instead it adds a special
case for this particular case and then handles it correctly as I was
unable to modify the code to handle it correctly.
It is not a nice solution, but it works.