Fix to Read and Write Sets (#1678)
During my work on the [new map
fusion](https://github.com/spcl/dace/issues/1643) I discovered a bug in
`SDFGState._read_and_write_set()`.
Originally I solved it there, but it was decided to move it into its own
PR.


Lets look at the first, super silly example, that is not useful on its
own.
The main point here, is that the `data` attribute of the Memlet does not
refer to the source of the connection but of the destination.


![test_1](https://github.com/user-attachments/assets/740ee4fc-cfe5-4844-a999-e316cb8f9c16)


BTW: The Webviewer outputs something like `B[0] -> [0, 0]` however, the
parser of the Memlet constructor does not understand this, it must be
written as `B[0] -> 0, 0`, i.e. the second set of brackets must be
omitted, this should be changed!

From the above we would expect the following sets:
- Reads:
	- `A`: `[Range (0, 0)]`
- `B`: Should not be listed in this set, because it is fully read and
written, thus it is excluded.
- Writes
	- `B`: `[Range (0)]`
	- `C`: `[Range (0, 0), Range (1, 1)]`

However, the current implementation gives us:
- Reads: `{'A': [Range (0)], 'B': [Range (1, 1)]}`
- Write: `{'B': [Range (0)], 'C': [Range (1, 1), Range (0)]}`

The current behaviour is wrong because:
- `A` is a `2x2` array, thus the read set should also have two
dimensions.
- `B` inside the read set, it is a scalar, but the range has two
dimensions, furthermore, it is present at all.
- `C` the first member of the write set (`Range(1, 1)`) is correct,
while the second (`Range(0)`) is horrible wrong.


The second example is even more simple.


![test_2](https://github.com/user-attachments/assets/da3d03af-6f10-411f-952e-ab057ed057c6)


From the SDFG we expect the following sets:
- Reads:
	- `A`: `[Range(0, 0)]`
- Writes:
	- `B`: `[Range(0)]`

It is important that in the above example `other_subset` is `None` and
`data` is set to `A`, so it is not one of these "crazy" non standard
Memlets we have seen in the first test.
However, the current implementation gives us:
- Reads: `{'A': [Range (0, 0)]}`
- Writes: `{'B': [Range (0, 0)]}`

This clearly shows, that whatever the implementation does is not
correct.