Quality of live: Improved error messages (#1731)
Small quality of live PR that improves two error messages. Especially if
the state was called `state` (as per default), it was hard to parse at
first sight


![image](https://github.com/user-attachments/assets/500cf100-57b8-4140-b4f7-d6cb0091829b)


For some reason, `yapf` was also changing the formatting of other code
parts. I thus made two commits

1. Format the two files that I was about to edit.
2. Improve the error messages in the freshly formatted files.

This allows to separate real changes from purely stylistic ones.

Would it be an idea to enforce linting with a GitHub Actions workflow?
I'm happy to write an issue about this (and contribute early next year).

---------

Co-authored-by: Roman Cattaneo <>