Replace another FORTRAN test program with `gfortran -Wall` certified test program. (#1736)
A follow up of #1733 that took a bit longer to figure out.

This change breaks the test on `multi_sdfg` branch, but there the
front-end is doing something additionally wrong (it somehow promotes
`nclv` into a input parameter for the whole program) -- but it will even
take longer to figure out why.