Updated `Node.add_{in, out}_connector()`. (#1828)
Before every time the function was (shallow) copying the `dict` upon a
change. This was done to make use of the implicit casting functionality
of the `DictProperty`, however, it is not very efficient. This commit
changes the implementation of the function to perform the casting
explicitly, thus avoiding the need to copy the `dict` every time.
This will also increase the performance of the whole tool chain.