Symbol redeclaration fix (#1788)
If a symbol is passed as an argument to the SDFG and assigned on an
interstate edge, the framecode declares the scalar symbol again. This PR
fixes that issue.

The case: If there is an interstate assignment on a variable, `N=5`, but
if 'N` is also an input argument to the SDFG, then the generated code
does not compile because `N` is redeclared.