Added `FindSingleUseData` Pass (#1906)
Added `FindSingleUseData` analysis pass

This new pass scans the SDFG and computes the data that is only used in
one single place.
Essentially, this boils down to the check how often a data descriptor is
accessed
A data descriptor is classified if the following statements are true:
- There is exactly one AccessNode referring to the data descriptor.
- The data descriptor is not referred to on an interstate edge, nor in
the condition parts of `ConditionBlock`s or `LoopRegion`s.

This pass will be needed to speed up fusion passes.