Fix DSE pass for conditional block with single branch and unconditional execution (#1930)
This PR addresses #1928 by fixing some code in `DeadStateElimination`
(DSE) pass. This code was supposed to handle exactly the case reported
in the issue, but the check for single branch with unconditional
execution was not entirely correct. Test case added.


Includes a change in fpga tests to address a CI error not related to
this change:

https://github.com/spcl/dace/actions/runs/13242670532/job/36961472662?pr=1930