Enforce consistent formatting on CI (#1957)
## Description

As discussed in the weekly DaCe meeting today, we'd like to go ahead
with enforcing consistent formatting as part of CI. As a reminder, this
started with a discussion[^1] in December and there was no pushback on
the discussion page and also not in the meeting today.

We decided to proceed in the following way.

1. @tbennun will run `pre-commit` / `yapf` locally on the codebase and
push that, see https://github.com/spcl/dace/pull/1966
2. This PR wil then be merged afterwards and just contain the changes
needed to add the new CI job.

<!--
Tal, you might wanna a couple manual guards in places like


https://github.com/spcl/dace/blob/6a485cf310940d3299ca48bcc4357d207ae67d04/tests/codegen/codegen_used_symbols_test.py#L17-L30

and other outlined below to keep beautiful manual formatting that makes
sense in some (very limited) cases.
-->

[^1]: https://github.com/spcl/dace/discussions/1804

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>
Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>