Modified GPU Launches (#1968)
This commit partially addresses
[issue#1967](https://github.com/spcl/dace/issues/1967), in the sense
that it turns GPU kernels into a no-ops (i.e. does not start it at all)
if a dimension of the iteration space has zero or negative size.

The code generator contained such a check, which was called "empty
grid".
Thus in a sense this commit only clarifies the notion of what an empty
grid is.
Before this was only detected if the size was dependent on runtime
values, but not if the size was known at compilation time.
As explained in the issue, this leads to strange behaviour, for example
the map range `{"__i": "a:b", "__j": "10:0"}` the code generator will
add checks for `__i` but not for `__j`.
This was inconsistent with CPU backend.